db connection issue resolved. schema/seed files and scripts finalized
