+++ db/schema.sql
@@ -4,27 +4,27 @@ CREATE DATABASE company_db;
 USE company_db;
 
 CREATE TABLE department (
-    -- directions make no reference to not null/auto increment but will be adding --
     id INT NOT NULL AUTO_INCREMENT PRIMARY KEY,
-    name VARCHAR(30) NOT NULL
+    dept_name VARCHAR(30) NOT NULL
 );
 
 CREATE TABLE role (
     id INT NOT NULL AUTO_INCREMENT PRIMARY KEY,
     title VARCHAR(30) NOT NULL,
-    salary DECIMAL,
+    salary DECIMAL(6, 0),
+    department_id INT,
     FOREIGN KEY (department_id)
     REFERENCES department(id)
     ON DELETE SET NULL
-)
+);
 
 CREATE TABLE employee (
     id INT NOT NULL AUTO_INCREMENT PRIMARY KEY,
     first_name VARCHAR(30),
     last_name VARCHAR(30),
+    role_id INT,
+    manager_id INT,
     FOREIGN KEY (role_id)
     REFERENCES role(id)
     ON DELETE SET NULL
-    -- possible foreign key(?)
-    manager_id INT
-)
+);
